fix: resolve TypeScript compilation errors and create working simplified parser
- Remove problematic enhanced-parser.ts with complex type issues
- Create enhanced-parser-simple.ts with working functionality
- Fix type mismatches in components and index files
- Update imports to use simplified parser version
- Successfully compile and build the project
