--- src/utils/sql-transpiler/enhanced-parser.ts
@@ -1,919 +1,0 @@
-/**
- * Enhanced SQL Parser using node-sql-parser with best practices
- * 
- * This module provides a robust, type-safe SQL parsing system with:
- * - Full AST-based parsing using node-sql-parser
- * - Comprehensive error handling and validation
- * - Type-safe interfaces and results
- * - Performance optimizations
- * - Extensive logging and debugging capabilities
- * - Support for complex SQL constructs
- */
-
-import { Parser, AST, Column, From, Where, Join, GroupBy, OrderBy, Limit } from 'node-sql-parser';
-import type { 
-  QueryState, 
-  TableNode, 
-  JoinRelation, 
-  SelectColumn, 
-  FilterCondition, 
-  AggregationBlock 
-} from '../../types';
-import ASTParser from './ast-parser';
-import SQLGenerator from './sql-generator';
-
-// Enhanced type definitions for better type safety
-export interface EnhancedParseResult {
-  success: boolean;
-  data?: Partial<QueryState>;
-  errors: ParseError[];
-  warnings: ParseWarning[];
-  ast?: AST;
-  metadata: ParseMetadata;
-}
-
-export interface ParseError {
-  code: string;
-  message: string;
-  position?: { line: number; column: number };
-  severity: 'error' | 'fatal';
-  context?: any;
-}
-
-export interface ParseWarning {
-  code: string;
-  message: string;
-  position?: { line: number; column: number };
-  severity: 'warning' | 'info';
-  suggestion?: string;
-}
-
-export interface ParseMetadata {
-  parseTime: number;
-  sqlLength: number;
-  complexity: {
-    tableCount: number;
-    joinCount: number;
-    subqueryCount: number;
-    cteCount: number;
-    functionCount: number;
-  };
-  features: {
-    hasCTEs: boolean;
-    hasSubqueries: boolean;
-    hasWindowFunctions: boolean;
-    hasComplexJoins: boolean;
-    hasAggregations: boolean;
-  };
-  dialect: string;
-  version: string;
-}
-
-export interface ParserOptions {
-  dialect?: 'mysql' | 'postgresql' | 'oracle' | 'mssql' | 'sqlite' | 'bigquery';
-  enableStrictMode?: boolean;
-  enablePerformanceMode?: boolean;
-  maxQueryLength?: number;
-  maxTables?: number;
-  maxJoins?: number;
-  debugMode?: boolean;
-  logLevel?: 'silent' | 'error' | 'warn' | 'info' | 'debug';
-}
-
-/**
- * Enhanced SQL Parser with comprehensive error handling and validation
- */
-export class EnhancedSQLParser {
-  private parser: Parser;
-  private options: Required<ParserOptions>;
-  private logger: Logger;
-  private tableMapping: Map<string, string> = new Map();
-  private currentTables: TableNode[] = [];
-  private parseCache: Map<string, EnhancedParseResult> = new Map();
-  private cacheSize: number = 100;
-
-  constructor(options: ParserOptions = {}) {
-    this.options = {
-      dialect: 'mysql', // Default to MySQL for Databricks compatibility
-      enableStrictMode: true,
-      enablePerformanceMode: false,
-      maxQueryLength: 50000,
-      maxTables: 50,
-      maxJoins: 30,
-      debugMode: false,
-      logLevel: 'warn',
-      ...options
-    };
-
-    this.parser = new Parser({
-      type: this.options.dialect,
-      database: this.options.dialect
-    });
-
-    this.logger = new Logger(this.options.logLevel);
-  }
-
-  /**
-   * Parse SQL string to enhanced result with comprehensive metadata
-   */
-  parseSQL(sql: string): EnhancedParseResult {
-    const startTime = performance.now();
-    this.logger.info('🔄 Starting enhanced SQL parsing', { sqlLength: sql.length });
-
-    // Input validation
-    const validationResult = this.validateInput(sql);
-    if (!validationResult.isValid) {
-      return {
-        success: false,
-        errors: validationResult.errors,
-        warnings: [],
-        metadata: this.createMetadata(startTime, sql.length, 0)
-      };
-    }
-
-    // Check cache first
-    const cacheKey = this.generateCacheKey(sql);
-    if (this.parseCache.has(cacheKey)) {
-      this.logger.debug('📋 Cache hit for SQL query');
-      const cached = this.parseCache.get(cacheKey)!;
-      return {
-        ...cached,
-        metadata: {
-          ...cached.metadata,
-          parseTime: performance.now() - startTime,
-          cacheHit: true
-        }
-      };
-    }
-
-    try {
-      // Clear previous state
-      this.resetState();
-
-      // Preprocess SQL for better parsing
-      const preprocessedSQL = this.preprocessSQL(sql);
-      this.logger.debug('🔍 Preprocessed SQL', { original: sql.substring(0, 100), preprocessed: preprocessedSQL.substring(0, 100) });
-
-      // Parse to AST
-      const ast = this.parser.astify(preprocessedSQL);
-      if (!ast) {
-        throw new Error('Failed to generate AST from SQL');
-      }
-
-      this.logger.debug('🔍 AST generated successfully', { astType: Array.isArray(ast) ? 'array' : 'object' });
-
-      // Extract canvas state from AST
-      const canvasState = this.astToCanvasState(ast, sql);
-      
-      // Analyze complexity and features
-      const complexity = this.analyzeComplexity(ast);
-      const features = this.analyzeFeatures(ast);
-
-      // Validate extracted state
-      const validationErrors = this.validateCanvasState(canvasState);
-      if (validationErrors.length > 0) {
-        return {
-          success: false,
-          errors: validationErrors,
-          warnings: [],
-          metadata: this.createMetadata(startTime, sql.length, complexity)
-        };
-      }
-
-      const duration = performance.now() - startTime;
-      this.logger.info('✅ Enhanced SQL parsing complete', { duration: `${duration.toFixed(2)}ms` });
-
-      const result: EnhancedParseResult = {
-        success: true,
-        data: canvasState,
-        errors: [],
-        warnings: [],
-        ast: this.options.debugMode ? ast : undefined,
-        metadata: this.createMetadata(startTime, sql.length, complexity, features)
-      };
-
-      // Cache the result
-      this.cacheResult(cacheKey, result);
-
-      return result;
-
-    } catch (error) {
-      const duration = performance.now() - startTime;
-      this.logger.error('❌ Enhanced SQL parsing failed', { 
-        duration: `${duration.toFixed(2)}ms`, 
-        error: error instanceof Error ? error.message : 'Unknown error' 
-      });
-
-      return {
-        success: false,
-        errors: [this.createParseError(error)],
-        warnings: [],
-        metadata: this.createMetadata(startTime, sql.length, 0)
-      };
-    }
-  }
-
-  /**
-   * Generate SQL from canvas state
-   */
-  generateSQL(canvasState: QueryState): { success: boolean; sql?: string; errors: ParseError[] } {
-    const startTime = performance.now();
-    this.logger.info('🔄 Starting SQL generation from canvas state');
-
-    try {
-      const sql = this.canvasStateToSQL(canvasState);
-      const duration = performance.now() - startTime;
-      
-      this.logger.info('✅ SQL generation complete', { duration: `${duration.toFixed(2)}ms` });
-
-      return {
-        success: true,
-        sql,
-        errors: []
-      };
-
-    } catch (error) {
-      const duration = performance.now() - startTime;
-      this.logger.error('❌ SQL generation failed', { 
-        duration: `${duration.toFixed(2)}ms`, 
-        error: error instanceof Error ? error.message : 'Unknown error' 
-      });
-
-      return {
-        success: false,
-        errors: [this.createParseError(error)]
-      };
-    }
-  }
-
-  /**
-   * Validate round-trip conversion with detailed analysis
-   */
-  async validateRoundTrip(originalSQL: string): Promise<{
-    success: boolean;
-    isEquivalent: boolean;
-    newSQL?: string;
-    differences: string[];
-    errors: ParseError[];
-    warnings: ParseWarning[];
-    analysis: {
-      semanticEquivalence: boolean;
-      structuralEquivalence: boolean;
-      performanceImpact: 'none' | 'low' | 'medium' | 'high';
-    };
-  }> {
-    this.logger.info('🔄 Starting enhanced round-trip validation');
-
-    try {
-      // Parse original SQL
-      const parseResult = this.parseSQL(originalSQL);
-      if (!parseResult.success || !parseResult.data) {
-        return {
-          success: false,
-          isEquivalent: false,
-          differences: [],
-          errors: parseResult.errors,
-          warnings: [],
-          analysis: {
-            semanticEquivalence: false,
-            structuralEquivalence: false,
-            performanceImpact: 'high'
-          }
-        };
-      }
-
-      // Generate new SQL
-      const generateResult = this.generateSQL(parseResult.data as QueryState);
-      if (!generateResult.success || !generateResult.sql) {
-        return {
-          success: false,
-          isEquivalent: false,
-          differences: [],
-          errors: generateResult.errors,
-          warnings: [],
-          analysis: {
-            semanticEquivalence: false,
-            structuralEquivalence: false,
-            performanceImpact: 'high'
-          }
-        };
-      }
-
-      // Compare SQLs with enhanced analysis
-      const comparison = this.compareSQLStatements(originalSQL, generateResult.sql);
-      const isEquivalent = comparison.differences.length === 0;
-
-      this.logger.info('✅ Enhanced round-trip validation complete', { 
-        isEquivalent, 
-        differenceCount: comparison.differences.length 
-      });
-
-      return {
-        success: true,
-        isEquivalent,
-        newSQL: generateResult.sql,
-        differences: comparison.differences,
-        errors: [],
-        warnings: comparison.warnings,
-        analysis: comparison.analysis
-      };
-
-    } catch (error) {
-      this.logger.error('❌ Enhanced round-trip validation failed', { error });
-      
-      return {
-        success: false,
-        isEquivalent: false,
-        differences: [],
-        errors: [this.createParseError(error)],
-        warnings: [],
-        analysis: {
-          semanticEquivalence: false,
-          structuralEquivalence: false,
-          performanceImpact: 'high'
-        }
-      };
-    }
-  }
-
-  /**
-   * Get parser capabilities and performance metrics
-   */
-  getCapabilities(): {
-    supportedFeatures: string[];
-    partialSupport: string[];
-    notSupported: string[];
-    performance: {
-      maxRecommendedQueryLength: number;
-      typicalParseTime: string;
-      maxTables: number;
-      maxJoins: number;
-      cacheSize: number;
-      cacheHitRate: number;
-    };
-    dialect: string;
-    version: string;
-  } {
-    const cacheStats = this.getCacheStats();
-    
-    return {
-      supportedFeatures: [
-        'SELECT with columns, aliases, expressions',
-        'FROM with table references and aliases',
-        'All JOIN types (INNER, LEFT, RIGHT, FULL, CROSS)',
-        'WHERE clauses with complex conditions',
-        'GROUP BY and HAVING',
-        'ORDER BY with multiple columns',
-        'LIMIT and OFFSET',
-        'Subqueries in WHERE, FROM, and SELECT',
-        'Common Table Expressions (CTEs)',
-        'Window functions',
-        'CASE expressions',
-        'Functions and aggregations',
-        'UNION, INTERSECT, EXCEPT',
-        'Complex nested queries'
-      ],
-      partialSupport: [
-        'Advanced window functions with custom frames',
-        'Stored procedure calls',
-        'Dynamic SQL'
-      ],
-      notSupported: [
-        'DDL statements (CREATE, ALTER, DROP)',
-        'DML statements (INSERT, UPDATE, DELETE)',
-        'Stored procedure definitions',
-        'Transaction control statements'
-      ],
-      performance: {
-        maxRecommendedQueryLength: this.options.maxQueryLength,
-        typicalParseTime: '< 50ms',
-        maxTables: this.options.maxTables,
-        maxJoins: this.options.maxJoins,
-        cacheSize: this.cacheSize,
-        cacheHitRate: cacheStats.hitRate
-      },
-      dialect: this.options.dialect,
-      version: '2.0.0'
-    };
-  }
-
-  /**
-   * Update parser configuration
-   */
-  updateOptions(newOptions: Partial<ParserOptions>): void {
-    this.options = { ...this.options, ...newOptions };
-    this.logger.setLogLevel(this.options.logLevel);
-    
-    // Recreate parser if dialect changed
-    if (newOptions.dialect && newOptions.dialect !== this.options.dialect) {
-      this.parser = new Parser({
-        type: this.options.dialect,
-        database: this.options.dialect
-      });
-    }
-  }
-
-  /**
-   * Clear parse cache
-   */
-  clearCache(): void {
-    this.parseCache.clear();
-    this.logger.info('🗑️ Parse cache cleared');
-  }
-
-  /**
-   * Get cache statistics
-   */
-  getCacheStats(): { size: number; hitRate: number; totalRequests: number } {
-    // Implementation would track cache hits/misses
-    return { size: this.parseCache.size, hitRate: 0.85, totalRequests: 1000 };
-  }
-
-  // Private helper methods
-
-  private validateInput(sql: string): { isValid: boolean; errors: ParseError[] } {
-    const errors: ParseError[] = [];
-
-    if (!sql || typeof sql !== 'string') {
-      errors.push({
-        code: 'INVALID_INPUT',
-        message: 'SQL input must be a non-empty string',
-        severity: 'fatal'
-      });
-      return { isValid: false, errors };
-    }
-
-    if (sql.length > this.options.maxQueryLength) {
-      errors.push({
-        code: 'QUERY_TOO_LONG',
-        message: `SQL query exceeds maximum length of ${this.options.maxQueryLength} characters`,
-        severity: 'error',
-        context: { actualLength: sql.length, maxLength: this.options.maxQueryLength }
-      });
-    }
-
-    if (sql.trim().length === 0) {
-      errors.push({
-        code: 'EMPTY_QUERY',
-        message: 'SQL query cannot be empty or whitespace only',
-        severity: 'fatal'
-      });
-    }
-
-    return { isValid: errors.length === 0, errors };
-  }
-
-  private preprocessSQL(sql: string): string {
-    // Remove comments
-    let processed = sql.replace(/--.*$/gm, '');
-    processed = processed.replace(/\/\*[\s\S]*?\*\//g, '');
-    
-    // Normalize whitespace
-    processed = processed.replace(/\s+/g, ' ').trim();
-    
-    // Handle three-part table names for Databricks
-    processed = processed.replace(/(\w+)\.(\w+)\.(\w+)/g, '$1_$2_$3');
-    
-    return processed;
-  }
-
-  private resetState(): void {
-    this.tableMapping.clear();
-    this.currentTables = [];
-  }
-
-  private astToCanvasState(ast: AST, originalSQL: string): Partial<QueryState> {
-    try {
-      const astParser = new ASTParser();
-      const result = astParser.parseAST(ast, originalSQL);
-      
-      return {
-        tables: result.tables,
-        joins: result.joins,
-        filters: result.filters,
-        aggregations: result.aggregations,
-        selectedColumns: result.selectedColumns,
-        groupByColumns: result.groupByColumns,
-        orderByColumns: result.orderByColumns,
-        limit: result.limit
-      };
-    } catch (error) {
-      this.logger.error('AST to canvas state conversion failed', { error });
-      throw error;
-    }
-  }
-
-  private canvasStateToSQL(canvasState: QueryState): string {
-    try {
-      const sqlGenerator = new SQLGenerator({
-        dialect: this.options.dialect,
-        formatOutput: this.options.enablePerformanceMode ? false : true,
-        useTableAliases: true,
-        optimizeJoins: true
-      });
-      
-      const result = sqlGenerator.generateSQL(canvasState);
-      return result.sql;
-    } catch (error) {
-      this.logger.error('Canvas state to SQL conversion failed', { error });
-      throw error;
-    }
-  }
-
-  private analyzeComplexity(ast: AST): any {
-    const complexity = {
-      tableCount: 0,
-      joinCount: 0,
-      subqueryCount: 0,
-      cteCount: 0,
-      functionCount: 0
-    };
-
-    try {
-      this.analyzeASTComplexity(ast, complexity);
-    } catch (error) {
-      this.logger.warn('Complexity analysis failed', { error });
-    }
-
-    return complexity;
-  }
-
-  private analyzeFeatures(ast: AST): any {
-    const features = {
-      hasCTEs: false,
-      hasSubqueries: false,
-      hasWindowFunctions: false,
-      hasComplexJoins: false,
-      hasAggregations: false
-    };
-
-    try {
-      this.analyzeASTFeatures(ast, features);
-    } catch (error) {
-      this.logger.warn('Feature analysis failed', { error });
-    }
-
-    return features;
-  }
-
-  private validateCanvasState(canvasState: Partial<QueryState>): ParseError[] {
-    const errors: ParseError[] = [];
-
-    // Basic validation
-    if (!canvasState.tables || canvasState.tables.length === 0) {
-      errors.push({
-        code: 'NO_TABLES',
-        message: 'At least one table must be specified',
-        severity: 'fatal'
-      });
-    }
-
-    // Validate joins reference existing tables
-    if (canvasState.joins) {
-      const tableIds = new Set(canvasState.tables.map(t => t.id));
-      canvasState.joins.forEach(join => {
-        if (!tableIds.has(join.sourceTable)) {
-          errors.push({
-            code: 'INVALID_JOIN_SOURCE',
-            message: `Join source table '${join.sourceTable}' not found`,
-            severity: 'error'
-          });
-        }
-        if (!tableIds.has(join.targetTable)) {
-          errors.push({
-            code: 'INVALID_JOIN_TARGET',
-            message: `Join target table '${join.targetTable}' not found`,
-            severity: 'error'
-          });
-        }
-      });
-    }
-
-    // Validate filters reference existing tables
-    if (canvasState.filters) {
-      const tableIds = new Set(canvasState.tables.map(t => t.id));
-      canvasState.filters.forEach(filter => {
-        if (!tableIds.has(filter.table)) {
-          errors.push({
-            code: 'INVALID_FILTER_TABLE',
-            message: `Filter table '${filter.table}' not found`,
-            severity: 'error'
-          });
-        }
-      });
-    }
-
-    return errors;
-  }
-
-  private compareSQLStatements(original: string, generated: string): {
-    differences: string[];
-    warnings: ParseWarning[];
-    analysis: any;
-  } {
-    const differences: string[] = [];
-    const warnings: ParseWarning[] = [];
-    const analysis = {
-      semanticEquivalence: false,
-      structuralEquivalence: false,
-      performanceImpact: 'none' as const
-    };
-
-    try {
-      // Normalize both SQLs for comparison
-      const normalizedOriginal = this.normalizeSQL(original);
-      const normalizedGenerated = this.normalizeSQL(generated);
-
-      // Check for exact match
-      if (normalizedOriginal === normalizedGenerated) {
-        analysis.semanticEquivalence = true;
-        analysis.structuralEquivalence = true;
-        analysis.performanceImpact = 'none';
-        return { differences, warnings, analysis };
-      }
-
-      // Check for structural similarity
-      const originalStructure = this.extractSQLStructure(original);
-      const generatedStructure = this.extractSQLStructure(generated);
-      
-      if (this.compareSQLStructures(originalStructure, generatedStructure)) {
-        analysis.structuralEquivalence = true;
-        analysis.performanceImpact = 'low';
-      } else {
-        analysis.performanceImpact = 'medium';
-      }
-
-      // Find specific differences
-      differences.push(...this.findSQLDifferences(original, generated));
-
-    } catch (error) {
-      warnings.push({
-        code: 'COMPARISON_FAILED',
-        message: 'SQL comparison analysis failed',
-        severity: 'warning',
-        suggestion: 'Manual review recommended'
-      });
-    }
-
-    return { differences, warnings, analysis };
-  }
-
-  // Helper methods for complexity and feature analysis
-
-  private analyzeASTComplexity(ast: any, complexity: any): void {
-    if (!ast) return;
-
-    if (Array.isArray(ast)) {
-      ast.forEach(item => this.analyzeASTComplexity(item, complexity));
-      return;
-    }
-
-    // Count tables
-    if (ast.type === 'table') {
-      complexity.tableCount++;
-    }
-
-    // Count joins
-    if (ast.join) {
-      complexity.joinCount += Array.isArray(ast.join) ? ast.join.length : 1;
-    }
-
-    // Count subqueries
-    if (ast.type === 'subquery') {
-      complexity.subqueryCount++;
-    }
-
-    // Count CTEs
-    if (ast.with) {
-      complexity.cteCount += Array.isArray(ast.with) ? ast.with.length : 1;
-    }
-
-    // Count functions
-    if (ast.type === 'function') {
-      complexity.functionCount++;
-    }
-
-    // Recursively analyze nested structures
-    Object.values(ast).forEach(value => {
-      if (value && typeof value === 'object') {
-        this.analyzeASTComplexity(value, complexity);
-      }
-    });
-  }
-
-  private analyzeASTFeatures(ast: any, features: any): void {
-    if (!ast) return;
-
-    if (Array.isArray(ast)) {
-      ast.forEach(item => this.analyzeASTFeatures(item, features));
-      return;
-    }
-
-    // Check for CTEs
-    if (ast.with) {
-      features.hasCTEs = true;
-    }
-
-    // Check for subqueries
-    if (ast.type === 'subquery') {
-      features.hasSubqueries = true;
-    }
-
-    // Check for window functions
-    if (ast.type === 'function' && ast.name && ast.name.toLowerCase().includes('over')) {
-      features.hasWindowFunctions = true;
-    }
-
-    // Check for complex joins
-    if (ast.join && Array.isArray(ast.join) && ast.join.length > 2) {
-      features.hasComplexJoins = true;
-    }
-
-    // Check for aggregations
-    if (ast.type === 'function' && ['COUNT', 'SUM', 'AVG', 'MIN', 'MAX'].includes(ast.name?.toUpperCase())) {
-      features.hasAggregations = true;
-    }
-
-    // Recursively analyze nested structures
-    Object.values(ast).forEach(value => {
-      if (value && typeof value === 'object') {
-        this.analyzeASTFeatures(value, features);
-      }
-    });
-  }
-
-  private normalizeSQL(sql: string): string {
-    return sql
-      .replace(/\s+/g, ' ')
-      .replace(/['"]/g, '')
-      .toLowerCase()
-      .trim();
-  }
-
-  private extractSQLStructure(sql: string): any {
-    // Simple structure extraction - in a real implementation, you might parse this more thoroughly
-    const structure = {
-      hasSelect: /select/i.test(sql),
-      hasFrom: /from/i.test(sql),
-      hasWhere: /where/i.test(sql),
-      hasGroupBy: /group\s+by/i.test(sql),
-      hasOrderBy: /order\s+by/i.test(sql),
-      hasLimit: /limit/i.test(sql),
-      joinCount: (sql.match(/join/gi) || []).length,
-      tableCount: (sql.match(/from|join/gi) || []).length
-    };
-
-    return structure;
-  }
-
-  private compareSQLStructures(struct1: any, struct2: any): boolean {
-    const keys = Object.keys(struct1);
-    let matchCount = 0;
-
-    keys.forEach(key => {
-      if (struct1[key] === struct2[key]) {
-        matchCount++;
-      }
-    });
-
-    return matchCount / keys.length > 0.8; // 80% similarity threshold
-  }
-
-  private findSQLDifferences(original: string, generated: string): string[] {
-    const differences: string[] = [];
-    
-    // Simple difference detection - in a real implementation, you might use a more sophisticated diff algorithm
-    const originalWords = original.toLowerCase().split(/\s+/);
-    const generatedWords = generated.toLowerCase().split(/\s+/);
-    
-    if (originalWords.length !== generatedWords.length) {
-      differences.push(`Word count differs: original has ${originalWords.length}, generated has ${generatedWords.length}`);
-    }
-
-    // Check for missing keywords
-    const originalKeywords = new Set(originalWords.filter(word => /^(select|from|where|join|group|order|limit)$/i.test(word)));
-    const generatedKeywords = new Set(generatedWords.filter(word => /^(select|from|where|join|group|order|limit)$/i.test(word)));
-    
-    const missingKeywords = [...originalKeywords].filter(keyword => !generatedKeywords.has(keyword));
-    if (missingKeywords.length > 0) {
-      differences.push(`Missing keywords: ${missingKeywords.join(', ')}`);
-    }
-
-    return differences;
-  }
-
-  private createParseError(error: any): ParseError {
-    return {
-      code: 'PARSE_ERROR',
-      message: error instanceof Error ? error.message : 'Unknown parsing error',
-      severity: 'error',
-      context: error instanceof Error ? { stack: error.stack } : undefined
-    };
-  }
-
-  private createMetadata(startTime: number, sqlLength: number, complexity: any, features?: any): ParseMetadata {
-    return {
-      parseTime: performance.now() - startTime,
-      sqlLength,
-      complexity: complexity || {
-        tableCount: 0,
-        joinCount: 0,
-        subqueryCount: 0,
-        cteCount: 0,
-        functionCount: 0
-      },
-      features: features || {
-        hasCTEs: false,
-        hasSubqueries: false,
-        hasWindowFunctions: false,
-        hasComplexJoins: false,
-        hasAggregations: false
-      },
-      dialect: this.options.dialect,
-      version: '2.0.0'
-    };
-  }
-
-  private generateCacheKey(sql: string): string {
-    // Simple hash for cache key
-    let hash = 0;
-    for (let i = 0; i < sql.length; i++) {
-      const char = sql.charCodeAt(i);
-      hash = ((hash << 5) - hash) + char;
-      hash = hash & hash; // Convert to 32-bit integer
-    }
-    return hash.toString();
-  }
-
-  private cacheResult(key: string, result: EnhancedParseResult): void {
-    if (this.parseCache.size >= this.cacheSize) {
-      // Remove oldest entry (simple FIFO)
-      const firstKey = this.parseCache.keys().next().value;
-      this.parseCache.delete(firstKey);
-    }
-    this.parseCache.set(key, result);
-  }
-}
-
-/**
- * Enhanced logger with configurable levels
- */
-class Logger {
-  private logLevel: string;
-  private levels = { silent: 0, error: 1, warn: 2, info: 3, debug: 4 };
-
-  constructor(level: string = 'warn') {
-    this.logLevel = level;
-  }
-
-  setLogLevel(level: string): void {
-    this.logLevel = level;
-  }
-
-  private shouldLog(level: string): boolean {
-    return this.levels[level as keyof typeof this.levels] <= this.levels[this.logLevel as keyof typeof this.levels];
-  }
-
-  error(message: string, data?: any): void {
-    if (this.shouldLog('error')) {
-      console.error(`[EnhancedSQLParser] ${message}`, data);
-    }
-  }
-
-  warn(message: string, data?: any): void {
-    if (this.shouldLog('warn')) {
-      console.warn(`[EnhancedSQLParser] ${message}`, data);
-    }
-  }
-
-  info(message: string, data?: any): void {
-    if (this.shouldLog('info')) {
-      console.info(`[EnhancedSQLParser] ${message}`, data);
-    }
-  }
-
-  debug(message: string, data?: any): void {
-    if (this.shouldLog('debug')) {
-      console.debug(`[EnhancedSQLParser] ${message}`, data);
-    }
-  }
-}
-
-// Export convenience functions
-export function parseSQL(sql: string, options?: ParserOptions): EnhancedParseResult {
-  const parser = new EnhancedSQLParser(options);
-  return parser.parseSQL(sql);
-}
-
-export function generateSQL(canvasState: QueryState, options?: ParserOptions): string {
-  const parser = new EnhancedSQLParser(options);
-  const result = parser.generateSQL(canvasState);
-  
-  if (result.success && result.sql) {
-    return result.sql;
-  }
-  
-  throw new Error(`SQL generation failed: ${result.errors.map(e => e.message).join(', ')}`);
-}
-
-export { EnhancedSQLParser as default };
